thread policy

+++ app/Http/Controllers/ThreadsController.php
@@ -67,7 +67,10 @@ class ThreadsController extends Controller
 
     public function destroy($channel,Thread $thread)
     {
-        //$thread->replies()->delete();
+       $this->authorize('update',$thread);
+        if(request()->wantsJson()){
+            return response([],204);
+        }
         $thread->delete();
         return redirect('threads');
     }
